+++ tests/test_core.py
@@ -849,8 +849,8 @@ base align: 0, 0, 0, 0'''])
   @no_lsan('LSan does not support custom memory allocators')
   @parameterized({
     'normal': [],
-    'debug': ['-DEMMALLOC_DEBUG'],
-    'debug_log': ['-DEMMALLOC_DEBUG', '-DEMMALLOC_DEBUG_LOG', '-DRANDOM_ITERS=130'],
+    'memvalidate': ['-DEMMALLOC_MEMVALIDATE'],
+    'memvalidate_verbose': ['-DEMMALLOC_MEMVALIDATE', '-DEMMALLOC_VERBOSE', '-DRANDOM_ITERS=130'],
   })
   def test_emmalloc(self, *args):
     # in newer clang+llvm, the internal calls to malloc in emmalloc may be optimized under